+++ src/components/Header/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Container } from "./styles";
+
+export default function Header() {
+  return <Container></Container>;
+}
