+++ src/components/Header/styles.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components/native";
+
+export const Container = styled.View`
+  flex: 1;
+`;
