+++ src/screens/OAuthSignIn/index.tsx
@@ -1,24 +1,35 @@
+import { useOAuth } from "@clerk/clerk-expo";
 import AppButton from "@components/AppButton";
 import AppScreenContainer from "@components/AppScreenContainer";
 import AppText from "@components/AppText";
-import { appUseAuth } from "@hooks/AuthContext";
+import { useWarmUpBrowser } from "@hooks/warmUpBrowser";
 import { useNavigation } from "@react-navigation/native";
 import { useState } from "react";
 import { ActivityIndicator } from "react-native";
 
 export default function OAuthSignIn() {
-  const navigation = useNavigation();
+  useWarmUpBrowser();
 
-  const { appSignIn } = appUseAuth();
+  const navigation = useNavigation();
 
   const [isLoading, setIsLoading] = useState(false);
 
-  async function handleLogin() {
-    setIsLoading(true);
-    await appSignIn();
+  const { startOAuthFlow } = useOAuth({ strategy: "oauth_google" });
 
-    setIsLoading(false);
-    navigation.goBack();
+  async function appSignIn() {
+    setIsLoading(true);
+    try {
+      const { createdSessionId, setActive } = await startOAuthFlow();
+
+      if (createdSessionId) {
+        setActive!({ session: createdSessionId });
+        navigation.goBack();
+      }
+    } catch (err) {
+      console.error("OAuth error", err);
+    } finally {
+      setIsLoading(false);
+    }
   }
 
   return (
@@ -27,7 +38,7 @@ export default function OAuthSignIn() {
       {isLoading ? (
         <ActivityIndicator />
       ) : (
-        <AppButton title="Google login" onPress={handleLogin} />
+        <AppButton title="Google login" onPress={appSignIn} />
       )}
     </AppScreenContainer>
   );
